commit #8

+++ app/models/Delete.php
@@ -15,35 +15,57 @@ class Delete {
         $db = new Database;
         $this->conn = $db->getConnection();
     }
-   
-    
 
 
 
- public function deleteUser(int $id){
+
+
+
+
+
+/**
+ * DELETE USER
+ * 
+ *
+ * @access public
+ * @param int $id user id
+ * @return int 
+ */
+public function deleteUser(int $id){
         
     $sql = 'DELETE FROM users WHERE id = :id';
     $stmt = $this->conn->prepare($sql); 
     $stmt->bindParam(':id', $id, PDO::PARAM_INT); 
     $stmt->execute(); 
-    return $stmt->rowCount();
- }
+    return $stmt->rowCount(); // integer
+}
 
 
 
 
 
-  /*********************************************************************************************************************|
-    * DELETE IMAGE                                                                                                      |
-    * quando vogliamo cambiare l'immagine di un post                                                                    |
-    * dobbiamo prima eliminare il file immagine attuale che sta nella cartella dove è immagazzinato,                    |
-    * Ci occorre il nome del'immagine il quale lo otteniamo facendo una select al database con l'id del post            |                 
-    * Utilizziamo con la funzione builtin di php 'unlink' passandoci il percorso del file più il nome del file          |
-    * l'immagine viene eliminata                                                                                        |
-    ********************************************************************************************************************/
-    public function deleteImage($id) {
+/**
+ * DELETE IMAGE  
+ * 
+ * Questo metodo si occupa solo di eliminare file dell'immagine vecchia
+ * 
+ * Quando si vuole cambiare l'immagine (in questo caso l' immagine dell' avatar)    
+ * dobbiamo prima eliminare il file immagine attuale che sta nella cartella: public/image/avatar/,  
+ * per dare il comando per cancellare il file esatto dobbiamo prima ottenere il nome del file,
+ * il nome del file è immagazzinato nel database, della tabella users, del campo img, della riga che ha il numero del id che abbiamo
+ * 
+ * una volta ottenuto il nome del file dell'immagine vecchia e se l'immagine vecchia è diversa dall' immagine di default
+ * l' immagine di default non deve essere mai cancellata 
+ * utilizzando la funzione builtin di php 'unlink()' passandoci come argomento il percorso del file più il nome del file    
+ * file dell'immagine vecchia viene eliminato   
+ * 
+ * @access public
+ * @global string IMAGE_DEFAULT
+ * @param int $id user id
+ * @return void
+ */
+    public function deleteImage(int $id) {
     
-        $id = (int)$id;
         $sql = "SELECT img FROM users WHERE id = :id";
 
         $stmt = $this->conn->prepare($sql);
@@ -55,26 +77,34 @@ class Delete {
             $user = $stmt->fetch(PDO::FETCH_OBJ);
 
             $stmt = null; 
-            
-            unlink("public/image/avatar/".$user->img); //elimina il file del'immagine attuale
-
-            $this->setImageDefault($id); // Setta il nome dell' immagine di default
+            if ( $user->img != IMAGE_DEFAULT ) { // 'avatar__default.jpg'
+                
+                unlink("public/image/avatar/".$user->img); //elimina il file del'immagine attuale
+            }
+           // $this->setImageDefault($id); // Setta il nome dell' immagine di default
             
         }
     }
 
 
 
-    
 
-/***************************************************************************************|
-* SET IMAGE DEFAULT                                                                                
-* Assegna immagine di default                                                 
-* 
-****************************************************************************************/
-private function setImageDefault(int $id){
+
+/**
+ * SET IMAGE DEFAULT
+ * 
+ * Questo metodo viene richiamato perchè è stata eliminata l' immagine di avatar di un user
+ * Quindi setta il nome dell' immagine di default all'inteno del database nella
+ * tabella users, nel campo img, della riga che ha il numero del id del parametro di questo metodo
+ *
+ * @access public
+ * @global string IMAGE_DEFAULT
+ * @param int $id user id
+ * @return void 
+ */
+public function setImageDefault(int $id){
         
-    $imageName = 'avatar__default.png';
+    $imageName = IMAGE_DEFAULT; // 'avatar__default.jpg';
                            
     $sql = "UPDATE users SET img = :img WHERE id = :id";
 
@@ -85,15 +115,12 @@ private function setImageDefault(int $id){
    
     $stmt->execute();
     $stmt = null;
-   
- }
-
-
-
+    }
 } // chiude classe
 
 
 /**
+ * echo '<br>';
  * die( $ );
  * die( '' );
  * var_dump( $ );
